Highlight marker for selected point list row in the GoogleMap.
The marker corresponding to the row selected in the point list of the assessment
tab is now highlighted in the GoogleMap. This works by passing a new
argument `markers` to the `setPreview` function. The regular blue
markers are used as default, and a new vector of markers is generated
using leaflet's `makeIcon` function when a row is selected.