Zoom function added to pointlist.
A zoom button is added to the pointlist. It is based on the _Superzip_
example from the shiny gallery. The underlying java code is stored in
in the `gomap.js` file and is referenced in the server script as an
additional column in the pointlist, each line containing a chunk of
HTML code of class `go-map` with longitude and latitude, which is passed to the
gomap button. The table is passed to `DT::datatable` as a json object via
`DT::dataTableAjax`. An observer reads latitude and longitude from
`input$gomap` when the button is clicked and updates the googlemap.
`ui.r` includes `gomap.js` as a `<div>` in the header. Some bugs remain
due to the different handling of options by `DT::datatable` and
`DT::dataTableAjax` -- line selection is not set to `single` and row
names cannot be excluded at the moment.