Added zoom to point and zoom to grid buttons, changed map update structure.
Zoom functions based on the `superZip` example from the
shiny gallery were removed in favor of a simpler method
based on a single 'zoom to point' button. The sampling
grid object is now saved to a global variable to make
it easier for multiple observers to access it. Consequently,
the `setPreview()` function was removed. An `udpateGoogleMap()`
function was added as workaround for a bug appearing in
updating the google map in the assessment tab.